Verify user token when deleting blogs

+++ part4/controllers/blogs.js
@@ -51,7 +51,39 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  // get blog
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(204).end()
+  }
+
+  // get request of user making the request
+  const decodedUserToken = jwt.verify(
+    request.token, 
+    process.env.SECRET)
+
+  if (!decodedUserToken) {
+    return response.status(401)
+      .json({ error: 'invalid token' })
+  }
+
+  // Get author of request from decoded token
+  const user = await User.findById(decodedUserToken.id)
+  if (!user) {
+    return response.status(400)
+      .json({ error: 'UserId missing or not valid' })
+  }
+
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(401)
+      .json({ error: 'only the creator can delete a blog' })
+  }
+
+  // Delete the blog from both Blog and User models
   await Blog.findByIdAndDelete(request.params.id)
+  user.blogs = user.blogs
+    .filter(blogId=> blogId.toString() !== request.params.id)
+  await user.save()
 
   response.status(204).end()
 })
